feat: Create and Add pagination

+++ src/products/products.service.ts
@@ -2,6 +2,8 @@ import { Injectable } from "@nestjs/common";
 import { CreateProductDto } from "./dto/create-product.dto";
 import { UpdateProductDto } from "./dto/update-product.dto";
 import { PrismaService } from "../prisma/prisma.service";
+import { PaginationDto } from "../common/dto";
+import { envs } from "../configs";
 
 @Injectable()
 export class ProductsService {
@@ -17,8 +19,13 @@ export class ProductsService {
     });
   }
 
-  async findAll() {
-    return this.prisma.product.findMany();
+  async findAll(paginationDto: PaginationDto) {
+    const { page = envs.DEFAULT_PAGE_SIZE, limit = envs.DEFAULT_LIMIT_PAGE } =
+      paginationDto;
+    return this.prisma.product.findMany({
+      skip: (page - 1) * limit,
+      take: limit,
+    });
   }
 
   findOne(idproduct: number) {
@@ -39,7 +46,13 @@ export class ProductsService {
   }
 
   remove(idproduct: number) {
-    return this.prisma.product.delete({
+    const product = this.prisma.product.findUnique({ where: { idproduct } });
+
+    if (product == null) {
+      throw new Error(`Product with id ${idproduct} not found`);
+    }
+
+    this.prisma.product.delete({
       where: { idproduct },
     });
   }
